cohorts/2025/project/requirements.txt
deepgram_sdk==4.8.1
groq==0.31.0
httpx==0.28.1
langchain_core==0.3.72
langchain_groq==0.3.7
langchain_openai==0.3.28
langgraph==0.6.3
numpy==2.3.2
openlit==1.35.0
pandas==2.3.1
pymongo==4.14.0
python-dotenv==1.1.1
scikit_learn==1.7.1
streamlit==1.48.0
streamlit_feedback==0.1.4
twikit==2.3.3
lancedb==0.24.2
sentence_transformers==5.1.0
tiktoken==0.10.0
